backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
tensorflow
numpy
python-dotenv